+++ OneBlink.SDK/models/FormMigrationEmbeddedForms.cs
@@ -1,0 +1,15 @@
+namespace OneBlink.SDK.Model
+{
+    public class FormMigrationEmbeddedForms
+    {
+        public string sourceElementId
+        {
+            get; set;
+        }
+
+        public long targetFormId
+        {
+            get; set;
+        }
+    }
+}